refactor: move source code to src folder
