+++ ETF.Meido.sln
@@ -13,13 +13,23 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "3 - Infrastructure Layer", 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "4 - Cross-Cutting Layer", "4 - Cross-Cutting Layer", "{E38EA358-01B1-408D-9FCB-57BD3F528F5F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETF.Meido.WebApi", "ETF.Meido.WebApi\ETF.Meido.WebApi.csproj", "{7E7644B0-7FB4-4EC0-AE52-6B25E79D8161}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETF.Meido.WebApi", "src\ETF.Meido.WebApi\ETF.Meido.WebApi.csproj", "{7E7644B0-7FB4-4EC0-AE52-6B25E79D8161}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETF.Meido.CrossCutting.DataTransferObject", "ETF.Meido.CrossCutting.DataTransferObject\ETF.Meido.CrossCutting.DataTransferObject.csproj", "{B36102EE-9480-4B80-A7A7-5B76DF0B9884}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETF.Meido.CrossCutting.DataTransferObject", "src\ETF.Meido.CrossCutting.DataTransferObject\ETF.Meido.CrossCutting.DataTransferObject.csproj", "{B36102EE-9480-4B80-A7A7-5B76DF0B9884}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETF.Meido.CrossCutting.Errors.Contracts", "ETF.Meido.CrossCutting.Error.Contracts\ETF.Meido.CrossCutting.Errors.Contracts.csproj", "{DA4D0246-1A7F-409E-B557-1E7A1F946D23}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETF.Meido.CrossCutting.Errors.Contracts", "src\ETF.Meido.CrossCutting.Error.Contracts\ETF.Meido.CrossCutting.Errors.Contracts.csproj", "{DA4D0246-1A7F-409E-B557-1E7A1F946D23}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETF.Meido.CrossCutting.Errors.Implementation", "ETF.Meido.CrossCutting.Error.Implementation\ETF.Meido.CrossCutting.Errors.Implementation.csproj", "{29494C70-A357-4781-87A4-203B37CEA24E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETF.Meido.CrossCutting.Errors.Implementation", "src\ETF.Meido.CrossCutting.Error.Implementation\ETF.Meido.CrossCutting.Errors.Implementation.csproj", "{29494C70-A357-4781-87A4-203B37CEA24E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETF.Meido.CrossCutting.Notifications", "src\ETF.Meido.CrossCutting.Notification\ETF.Meido.CrossCutting.Notifications.csproj", "{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{A54C3FA7-1D0D-419B-B114-A2E1ECB74F90}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "4 - Cross-Cutting Layer", "4 - Cross-Cutting Layer", "{7E4B192C-A7F9-4815-8698-5530E51EAC8E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETF.Meido.CrossCutting.Errors.UnitTests", "test\ETF.Meido.CrossCutting.Errors.UnitTests\ETF.Meido.CrossCutting.Errors.UnitTests.csproj", "{44DD850F-AA59-4F75-A2C7-459ADEDF4469}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,15 +53,31 @@ Global
 		{29494C70-A357-4781-87A4-203B37CEA24E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{29494C70-A357-4781-87A4-203B37CEA24E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{29494C70-A357-4781-87A4-203B37CEA24E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44DD850F-AA59-4F75-A2C7-459ADEDF4469}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44DD850F-AA59-4F75-A2C7-459ADEDF4469}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44DD850F-AA59-4F75-A2C7-459ADEDF4469}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44DD850F-AA59-4F75-A2C7-459ADEDF4469}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{85A36BEE-FF0D-4EB0-AFE1-3B5B81C3F814} = {EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}
+		{634214B2-37CD-407F-913F-D43D6265F3FD} = {EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}
+		{0E7FC9E7-FEA9-4741-B9F7-8D0C85820BCE} = {EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}
+		{C6A283F5-B7B0-4177-ABE6-4C7D0418BFC1} = {EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}
+		{E38EA358-01B1-408D-9FCB-57BD3F528F5F} = {EA7F58A8-A586-43DA-AE9D-94EEFDF68E9E}
 		{7E7644B0-7FB4-4EC0-AE52-6B25E79D8161} = {85A36BEE-FF0D-4EB0-AFE1-3B5B81C3F814}
 		{B36102EE-9480-4B80-A7A7-5B76DF0B9884} = {E38EA358-01B1-408D-9FCB-57BD3F528F5F}
 		{DA4D0246-1A7F-409E-B557-1E7A1F946D23} = {E38EA358-01B1-408D-9FCB-57BD3F528F5F}
 		{29494C70-A357-4781-87A4-203B37CEA24E} = {E38EA358-01B1-408D-9FCB-57BD3F528F5F}
+		{6B16C0E8-A2D2-42AD-847E-4C11130A7C6C} = {E38EA358-01B1-408D-9FCB-57BD3F528F5F}
+		{7E4B192C-A7F9-4815-8698-5530E51EAC8E} = {A54C3FA7-1D0D-419B-B114-A2E1ECB74F90}
+		{44DD850F-AA59-4F75-A2C7-459ADEDF4469} = {7E4B192C-A7F9-4815-8698-5530E51EAC8E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9A8827B2-0973-408D-BA5F-F15D6430511A}
