+++ src/ETF.Meido.CrossCutting.Error.Contracts/ErrorType.cs
